backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
langgraph==0.2.60
langchain-google-genai==2.0.8
pydantic==2.10.4
aiosqlite==0.20.0
python-dotenv==1.0.1
